brand_tool_demo/requirements.txt
Flask==2.3.3
google-genai==0.5.0
requests==2.31.0
Werkzeug==2.3.7 